agent/requirements.txt
copilotkit>=0.1.54
fastapi>=0.115.14
langchain>=0.3.26
langchain-openai>=0.3.27
langgraph>=0.4.10
uvicorn>=0.35.0
python-dotenv>=1.0.0